+++ saleor/checkout/tests/test_checkout.py
@@ -970,12 +970,11 @@ def test_recalculate_checkout_discount_expired_voucher(checkout_with_voucher, vo
 
 
 def test_recalculate_checkout_discount_free_shipping_subtotal_less_than_shipping(
-    checkout_with_voucher_percentage_and_shipping,
-    voucher_free_shipping,
+    checkout_with_voucher_free_shipping,
     shipping_method,
     channel_USD,
 ):
-    checkout = checkout_with_voucher_percentage_and_shipping
+    checkout = checkout_with_voucher_free_shipping
     manager = get_plugins_manager()
     lines, _ = fetch_checkout_lines(checkout)
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
@@ -987,14 +986,9 @@ def test_recalculate_checkout_discount_free_shipping_subtotal_less_than_shipping
         address=checkout.shipping_address,
     ).gross + Money("10.00", "USD")
     channel_listing.save()
-    checkout.price_expiration = timezone.now()
-    checkout.save()
 
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
     recalculate_checkout_discount(manager, checkout_info, lines, None)
-
-    checkout.price_expiration = timezone.now()
-    checkout.save()
 
     assert checkout.discount == channel_listing.price
     assert checkout.discount_name == "Free shipping"
@@ -1014,12 +1008,11 @@ def test_recalculate_checkout_discount_free_shipping_subtotal_less_than_shipping
 
 
 def test_recalculate_checkout_discount_free_shipping_subtotal_bigger_than_shipping(
-    checkout_with_voucher_percentage_and_shipping,
-    voucher_free_shipping,
+    checkout_with_voucher_free_shipping,
     shipping_method,
     channel_USD,
 ):
-    checkout = checkout_with_voucher_percentage_and_shipping
+    checkout = checkout_with_voucher_free_shipping
     manager = get_plugins_manager()
     lines, _ = fetch_checkout_lines(checkout)
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
@@ -1031,13 +1024,9 @@ def test_recalculate_checkout_discount_free_shipping_subtotal_bigger_than_shippi
         address=checkout.shipping_address,
     ).gross - Money("1.00", "USD")
     channel_listing.save()
-    checkout.price_expiration = timezone.now()
-    checkout.save()
 
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
     recalculate_checkout_discount(manager, checkout_info, lines, None)
-    checkout.price_expiration = timezone.now()
-    checkout.save()
 
     assert checkout.discount == channel_listing.price
     assert checkout.discount_name == "Free shipping"